Moved SBX container content hashing into a separate function
